backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
z3-solver==4.13.4.0
requests==2.32.3
python-dotenv==1.0.1
